+++ crates/holochain_types/src/dht_op/tests.rs
@@ -1,10 +1,10 @@
 use super::OpBasis;
+use crate::prelude::fixt::*;
 use crate::prelude::*;
 use ::fixt::prelude::*;
 use holo_hash::fixt::ActionHashFixturator;
 use holo_hash::*;
 use holochain_trace;
-use crate::prelude::fixt::*;
 use tracing::*;
 
 struct RecordTest {